Refactor projects, move types to separate file
