+++ hooks/useFetchProjects.ts
@@ -1,60 +1,8 @@
 import { useEffect, useState } from 'react';
 import useLocalStorage from './useLocalStorage';
-import { ProjectFilter } from '@/components/Projects/FiltersBar/FilterBtnsGroup';
+import { CachedProjects, IRepositoriesAPIResponse, ProjectFilter, RepoItem } from '@/types';
 
-export interface IRepositoriesAPIResponse {
-  items: RepoItem[];
-}
 
-export interface RepoItem {
-  name: string;
-  openGraphImageUrl: string;
-  collaborators: {
-    totalCount: number;
-  };
-  description: null | string;
-  url: string;
-  createdAt: string;
-  updatedAt: string;
-  stargazerCount: number;
-  languages: Languages;
-  contributors: Contributors;
-}
-
-export interface Contributors {
-  edges: ContributorsEdge[];
-}
-
-export interface ContributorsEdge {
-  node: PurpleNode;
-}
-
-export interface PurpleNode {
-  avatarUrl: string;
-  login: string;
-}
-
-export interface Languages {
-  edges: LanguagesEdge[];
-}
-
-export interface LanguagesEdge {
-  node: FluffyNode;
-}
-
-export interface FluffyNode {
-  name: string;
-}
-
-export interface IGithubProjectsMaakafResponse {
-  githubLink: string;
-  discordLink: string;
-}
-
-interface CachedProjects {
-  items: IRepositoriesAPIResponse;
-  cacheTimestamp: string;
-}
 
 const cacheKey = 'projectsData';
 const apiEndpoint = '/api/projects'; // The new API endpoint
